+++ src/ConnectionFactory.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace KaduDutra\PsrContainerDoctrine;
 
 use Doctrine\DBAL\Connection;
-use Doctrine\DBAL\Driver\PDOMySql\Driver as PdoMysqlDriver;
+use Doctrine\DBAL\Driver\PDO\MySQL\Driver as PdoMysqlDriver;
 use Doctrine\DBAL\DriverManager;
 use Doctrine\DBAL\Types\Type;
 use Psr\Container\ContainerInterface;
@@ -16,8 +16,7 @@ use function is_string;
  */
 final class ConnectionFactory extends AbstractFactory
 {
-    /** @var bool */
-    private static $areTypesRegistered = false;
+    private static bool $areTypesRegistered = false;
 
     /**
      * {@inheritdoc}