+++ test/ConnectionFactoryTest.php
@@ -5,8 +5,9 @@ declare(strict_types=1);
 namespace KaduDutraTest\PsrContainerDoctrine;
 
 use Doctrine\Common\EventManager;
-use Doctrine\DBAL\Driver\PDOMySql\Driver as PDOMySQLDriver;
-use Doctrine\DBAL\Driver\PDOSqlite\Driver as PDOSqliteDriver;
+use Doctrine\DBAL\Driver\AbstractMySQLDriver;
+use Doctrine\DBAL\Driver\PDO\MySQL\Driver as PDOMySQLDriver;
+use Doctrine\DBAL\Driver\PDO\SQLite\Driver as PDOSqliteDriver;
 use Doctrine\DBAL\Exception\ConnectionException;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Doctrine\DBAL\Types\BooleanType;
@@ -21,14 +22,11 @@ use function sprintf;
 
 final class ConnectionFactoryTest extends TestCase
 {
-    /** @var Configuration */
-    private $configuration;
+    private Configuration $configuration;
 
-    /** @var EventManager */
-    private $eventManger;
+    private EventManager $eventManger;
 
-    /** @var AbstractPlatform */
-    private $customPlatform;
+    private AbstractPlatform $customPlatform;
 
     public function setUp() : void
     {
@@ -68,7 +66,7 @@ final class ConnectionFactoryTest extends TestCase
             $factory($container);
         } catch (ConnectionException $e) {
             foreach ($e->getTrace() as $entry) {
-                if ($entry['class'] === PDOMySQLDriver::class) {
+                if ($entry['class'] === PDOMySQLDriver::class || $entry['class'] === AbstractMySQLDriver::class) {
                     /** @psalm-suppress InternalMethod @todo find a better way to add to assertion count... */
                     $this->addToAssertionCount(1);
 
@@ -91,6 +89,7 @@ final class ConnectionFactoryTest extends TestCase
 
         $this->assertSame($this->configuration, $connection->getConfiguration());
         $this->assertSame($this->eventManger, $connection->getEventManager());
+        /** @psalm-suppress InternalMethod */
         $this->assertSame([
             'driverClass' => PDOSqliteDriver::class,
             'wrapperClass' => null,
@@ -126,6 +125,7 @@ final class ConnectionFactoryTest extends TestCase
             'params' => ['username' => 'foo'],
         ]));
 
+        /** @psalm-suppress InternalMethod */
         $this->assertSame([
             'username' => 'foo',
             'driverClass' => PDOSqliteDriver::class,